requirements.txt
# Core dependencies
pandas>=2.0.0
numpy>=1.24.0,<2.0.0
python-dotenv==1.0.0
setuptools>=65.0.0

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.23

# Data validation
great-expectations==0.18.8

# Time series analysis
statsmodels==0.14.0
pmdarima==2.0.4

# Machine learning
scikit-learn==1.3.2
xgboost==2.0.3
# lightgbm==4.1.0  # Optional - has build issues on Python 3.12, not used in core functionality

# Visualization
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.18.0

# Dashboard
streamlit==1.29.0

# Workflow automation (optional - comment out if having install issues)
# Note: Prefect depends on pendulum which has Python 3.12 compatibility issues
# You can skip this if not using workflow automation
prefect>=2.14.0; python_version < "3.12"
prefect>=2.14.0; python_version >= "3.12" and platform_machine == "x86_64"

# Utilities
requests==2.31.0
openpyxl==3.1.2

# Testing
pytest==7.4.3
pytest-cov==4.1.0

# Code quality
black==23.12.1
flake8==6.1.0

